Refactor: Migrate from custom DAOs to Spring Data JPA/MongoDB

+++ src/main/java/com/blog/DataTransporter/Comment/ResponseCommentDTO.java
@@ -4,6 +4,7 @@ import com.blog.Model.Comment;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import org.springframework.data.relational.core.sql.In;
 
 @Schema(description = "Response data for a comment")
 public record ResponseCommentDTO(
@@ -12,13 +13,13 @@ public record ResponseCommentDTO(
     String id,
     @NotNull
     @Schema(description = "ID of the user who created the comment", example = "1")
-    Long userId,
+    Integer userId,
     @NotBlank
     @Schema(description = "Username of the commenter", example = "johndoe")
     String username,
     @NotNull
     @Schema(description = "ID of the post being commented on", example = "1")
-    Long postId,
+    Integer postId,
     @NotBlank
     @Schema(description = "Content of the comment", example = "Great post! Very informative.")
     String body
